Update USB-C_Power.ino
Updated logic to use PD0 as a non-active, default state. 

This fixes the persistent PDO profile issue, where the last loaded PDO config is present when re-connecting a USB-C PD source to the board. 

This change means that the board will default to no output - which is more suitable for this use case!